+++ changelog.md
@@ -8,6 +8,7 @@
 - Change recycling code to be a little less hacky (Malte)
 - Fix some hard incompatabilities with Optifine (Malte)
 - Fix engineer's crafting table losing its inventory (Malte)
+- Fix placement and colour of various "Inventory" texts (BluSunrize)
 
 ##### Version 1.18.2-8.0.0-146
 - Re-add compat for OpenComputers2 (Malte)